デモ作成

+++ DirectXGame/Engine/DirectXCore/Buffer/ConstantBuffer.cpp
@@ -31,7 +31,7 @@ void ConstantBuffer::Create(size_t size)
 
 	constantBufferView = {};
 	constantBufferView.BufferLocation = buffer->GetGPUVirtualAddress();
-	constantBufferView.SizeInBytes = resDesc.Width;
+	constantBufferView.SizeInBytes = static_cast<UINT>(resDesc.Width);
 
 	DirectXBase::GetInstance()->GetDescriptorHeap()->CreateCBV(constantBufferView);
