+++ DirectXGame/Engine/DirectXCore/WindowsApp.h
@@ -7,8 +7,8 @@ private:
 	WNDCLASSEX w{};
 	HWND hwnd;
 	MSG msg{};//メッセージ
-	float window_width = 1280;//横幅
-	float window_height = 720;//縦幅
+	float window_width = 1280.0f;//横幅
+	float window_height = 720.0f;//縦幅
 
 public:
 
@@ -18,15 +18,15 @@ public:
 	/// <param name="title"></param>
 	/// <param name="width">ウィンドウの横幅の大きさ</param>
 	/// <param name="height">ウィンドウの縦幅の大きさ</param>
-	void CreatWindow(const wchar_t* title = L"DirectXGame", const int width = 1280, const int height = 720);
+	void CreatWindow(const wchar_t* title = L"DirectXGame", const float width = 1280.0f, const float height = 720.0f);
 	bool MessageWindow();
 	void Break();
 
 	//ゲッター
 	WNDCLASSEX GetW();
 	HWND GetHwnd();
-	int GetWidth();
-	int GetHeight();
+	float GetWidth();
+	float GetHeight();
 
 	//セッター
 	void SetW(WNDCLASSEX W);
